Include CloudFormation parameter files
Parameter files are used to provide CloudFormation parameter values
on the command line in an easy way. They can contain sensitive data,
so care should be taken when committing them in a repository.

They are included at this stage to make it easy to keep different
parameter lists for different branches.